style: 抽出 CalcSign 为公共 CalcSign
