+++ src/constant/index.ts
@@ -1,0 +1,9 @@
+
+export enum CalcSign {
+    Add= "+",
+    Subtract= "-",
+    Multiply= "*",
+    Divide= "/",
+    Back= "Back",
+    C= "C",
+}
