+++ src/pages/count/constant.ts
@@ -68,13 +68,4 @@ const data = [
     ]
 ]
 
-export enum CalcSign {
-    Add= "+",
-    Subtract= "-",
-    Multiply= "*",
-    Divide= "/",
-    Back= "Back",
-    C= "C",
-}
-
 export default data
